Create getNodeDistance
Given the root of a Binary Tree  and 2 integers that represent the data values of any two TreeNodes present in the tree, write a method - getNodeDistance that returns the distance between the nodes. You can assume that the given keys exist in the tree. The distance between two nodes is defined as the minimum number of edges that must be traversed to travel between the two nodes.